feat: 抽出 PipelineController 管理运行时管线
将音频采集、VAD、ASR 队列线程和增量 ASR 协调迁移到新的 pipeline_controller.py。

LiveTranslateApp 现在通过 PipelineController 事件接收音频电平和 ASR 文本，只保留 ASR 调用包装、内存日志、翻译调度和 UI 更新职责。

ASR 切换和设置同步改为调用 PipelineController 的公开方法，避免 main.py 直接管理 capture/asr 线程、队列和 VAD 状态。

同步 CLAUDE.md 的线程模型说明，并补齐 control_panel.py 中已有 Path 使用缺失的导入。
